requirements.txt
chromadb==0.6.3
pdfminer.six==20250327
streamlit==1.44.1
python-dotenv==1.1.0
openai==1.73.0
tiktoken==0.9.0
python-docx==1.1.2
sentence-transformers==4.0.2
langchain-community==0.3.21
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain-openai==0.3.12
watchdog==6.0.0
pypdf==5.4.0
docx2txt==0.9
accelerate==1.6.0
openai==1.73.0
pysqlite3-binary==0.5.4
